refactors 'persona' to 'demon' in most p5 files
