+++ src/app/games/p5/p5-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core'
 import { Routes, RouterModule } from '@angular/router'
 
 import { P5WrapperComponent } from './components/p5-wrapper/p5-wrapper.component'
-import { P5PersonaListComponent } from './components/p5-persona-list/p5-persona-list.component'
-import { P5PersonaEntryComponent } from './components/p5-persona-entry.component'
+import { P5PersonaListComponent } from './components/p5-demon-list/p5-demon-list.component'
+import { P5PersonaEntryComponent } from './components/p5-demon-entry.component'
 import { P5SkillListComponent } from './components/p5-skill-list.component'
 import { P5FusionTableComponent } from './components/p5-fusion-table.component'
 import { P5DemonBuilderComponent } from './components/p5-demon-builder.component'
