+++ src/app/games/p5/p5.module.ts
@@ -9,8 +9,8 @@ import { P5RoutingModule } from './p5-routing.module'
 import { P5WrapperComponent } from './components/p5-wrapper/p5-wrapper.component'
 import { P5HeaderComponent } from './components/p5-header/p5-header.component'
 import { P5BookmarksComponent } from './components/p5-bookmarks/p5-bookmarks.component'
-import { P5PersonaListComponent } from './components/p5-persona-list/p5-persona-list.component'
-import { P5PersonaEntryComponent } from './components/p5-persona-entry.component'
+import { P5PersonaListComponent } from './components/p5-demon-list/p5-demon-list.component'
+import { P5PersonaEntryComponent } from './components/p5-demon-entry.component'
 import { P5SkillListComponent } from './components/p5-skill-list.component'
 import { P5FusionTableComponent } from './components/p5-fusion-table.component'
 import { P5DemonBuilderComponent } from './components/p5-demon-builder.component'
